Add XORG_DEFAULT_OPTIONS meta-macro.
Many modules use the same set of m4 macros. Additions to the macros require
updating the configure.ac for each module affected. In the case of e.g.
protos and libs this means updating every single module.

This patch adds a XORG_DEFAULT_OPTIONS meta-macro that uses the other macros
defined. Macros can be added to this meta-macro and modules immediately get
to make use of it - without changes required to configure.ac.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>