Add XORG_ENABLE_DOCS to control the building of documentation
This macro is intended for modules to control, in a single option,
the building of the documentaion. Using --enable-docs=no, all
documentation is skipped, regardless of the tool used to build it.

It should be used in conjunction with other XORG_WITH_* macros
which handle cases where some specific tool is not available.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>