XORG_WITH_GROFF: add HAVE_GROFF_HTML Automake conditional
Groff uses grohtml to generate html output format. This program, in turn,
uses a number of pnm* commands from the netpbm package, psselect
from the psutils package and the ghostscript package.

These are independently installed, so they could be missing.
A check is made to ensure those dependencies are installed.
If not, the makefile can use the conditional to supress the target.

Reviewed-by: Dan Nicholson <dbn.lists@gmail.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>