docs: Take optional argument to control defaults
The typical default of enabling documentation if the correct tool is
found is usually the right thing to do. However, some packages such as
Xlib have huge amounts of documentation that few people would need in
every output format. Allow packages to specify the default for each tool
by passing an argument to the macro.

Signed-off-by: Dan Nicholson <dbn.lists@gmail.com>
Reviewed-by: Gaetan Nadon <memsize@videotron.ca>
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>