XORG_WITH_XSLTPROC: add XSLT processor command line tool
The primary use will be for DocBook/XML cross links database generation.

XSLT (Extensible Stylesheet Language Transformations) is a declarative,
XML-based language used for the transformation of XML documents.
The xsltproc command line tool is for applying XSLT stylesheets
to XML documents.

It is used under the cover by xmlto to generate html files from DocBook/XML.
The XSLT processor is often used as a standalone tool for transformations.
It should not be assumed that it is used only to work with documnetation.

Tested-by: Matt Dew <marcoz@osource.org>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>