XORG_WITH_FOP: add minimum version checking support
Adding this feature is motivated by a bug in fop 0.95 where
fop crashes in the presence of some characters like the latin capital
sharp s. Fop version 1.0 works correctly.

This is the same feature found in XMLTO and the likes.
In the macro public interface there is a shift in parameter position
for the DEFAULT parameter. Doing development on older libraries
(point releases on older versions) will still work the same way as
the "no" value will be interpreted as the minimum version. It won't
be found, so fop will be disabled anyway.

Only libX11 has used XORG_WITH_FOP(no) and will be changed with a version
bump on util-macros.

Reviewed-by: Dan Nicholson <dbn.lists@gmail.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>