Separate unknown warning options by language
If XORG_COMPILER_FLAGS is called more than once with separate languages,
the unknown warning options used internally for unknown warning checking
will be set the first time and then the cached value will be used for
subsequent languages. This is a problem if the compilers differ between
the languages.

This change ensures that the unknown warning options are namespaced so
multiple XORG_COMPILER_FLAGS calls with different languages are checked
separately.

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Reviewed-by: Gaetan Nadon<memsize@videotron.ca>