Use the right variable names for the compiler flags in XORG_TESTSET_CFLAG
Whilst populating the xorg_cv_ cache entries, and the subsequent
xorg_testset_, we always include the FLAG_PREFIX in the name, to
distinguish between C and C++ iterations. However, in the body of the
test function, we dropped the FLAG_PREFIX expansion and so always used
the empty string and hence did not correctly detect incorrect flags.

The bug was introduced with
commit 51f69514e65637f575df901ef942515e70314f6a
Author: Chase Douglas <chase.douglas@canonical.com>
Date:   Mon Mar 12 11:49:28 2012 -0700

    Separate unknown warning options by language

Reported-and-tested-by: Sedat Dilek <sedat.dilek@gmail.com>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Sedat Dilek <sedat.dilek@gmail.com>
Cc: Chase Douglas <chase.douglas@canonical.com>
Cc: Gaetan Nadon<memsize@videotron.ca>
Acked-by: Alan Coopersmith <alan.coopersmith@oracle.com>