Bump minimum Autoconf required version to 2.62
The main motivation is to catch-up with the development reality and
allow use of features in Autoconf 2.62 as well as Automake 1.11.

As usual this means no features found only in versions above those
specified are allowed.

This is implementing the policy change which is described in:
http://www.x.org/wiki/Building_the_X_Window_System/?updated#index2h3

Discussion on xorg minimum autotools requirements:
http://lists.x.org/archives/xorg-devel/2013-October/038325.html

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>