XORG_PROG_RAWCPP: Add TRADITIONALCPPFLAGS
In some cases we may want to have -traditional for proper whitespace preserving
without -undef, as we actually want the system definitions to be in place
so we can #ifdef on them. IE in xinit various .cpp files contain things like
 #ifdef __APPLE__ and #ifdef __linux__

So this patch adds a TRADITIONALCPPFLAGS variable which contains just
-traditional where applicable without the other RAWCPPFLAGS for unsetting
the system definitions.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Reviewed-by: Gaetan Nadon <memsize@videotron.ca>